goforless/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
cssbeautifier==1.14.6
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Django==4.1.1
djlint==1.12.3
docopt==0.6.2
EditorConfig==0.12.3
entrypoints==0.4
executing==0.8.3
fonttools==4.33.3
frozenlist==1.3.0
gunicorn==20.1.0
h11==0.13.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.13.0
ipython==8.4.0
ischedule==1.2.5
jedi==0.18.1
Jinja2==3.1.2
jsbeautifier==1.14.6
jupyter-client==7.3.1
jupyter-core==4.10.0
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.22.4
outcome==1.1.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipreqs==0.4.11
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cord==2.0.0b7
pycparser==2.21
Pygments==2.12.0
pyodbc==4.0.32
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.0.0
regex==2022.8.17
requests==2.27.1
schedule==1.1.0
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.1
traitlets==5.2.1.post0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wcwidth==0.2.5
wsproto==1.1.0
whitenoise==6.2.0
yarg==0.1.9
yarl==1.7.2
zipp==3.8.1
